Refactored state system; no longer uses custom event.rs
This is a huge change and gets rid of a nasty code smell. Instead of using
a custom version of ggez's event module... we don't. I was not as experienced
at the time, and have learnt that the method I was using to create a state manager
and pass around state wasn't necessary. Now, there is only one "state" in the ggez
sense: the state manager. All other states are using a custom state system that works
within ggez's `EventHandler` trait instead of using a custom version of it.
