+++ src/states/menu_state/mod.rs
@@ -2,10 +2,11 @@ mod spawner;
 
 use std::time::Duration;
 use ggez::{Context, GameResult, graphics};
+use ggez::event::{Keycode, Mod, Button};
 use ggez::graphics::{Point, Color};
-use crate::event::{Assets, Transition, EventHandler, Keycode, Mod, Button};
 use crate::states::shared::option::{Option, OptionInputCommand};
 use crate::states::play_state::PlayState;
+use crate::states::{State, Assets, Transition};
 use crate::util::{DurationExt, play_click_sfx};
 use self::spawner::Spawner;
 
@@ -86,7 +87,7 @@ impl MenuState {
     }
 }
 
-impl EventHandler for MenuState {
+impl State for MenuState {
     fn update(
         &mut self,
         ctx: &mut Context,
