+++ src/states/menu_state/spawner.rs
@@ -1,10 +1,10 @@
-use std::time::Duration;
-use rand;
-use ggez::{Context, graphics};
-use ggez::graphics::Point;
-use crate::event::Assets;
-use crate::states::play_state::tetromino::{Piece, PieceType, u8_to_piece_type};
+use crate::states::play_state::tetromino::{u8_to_piece_type, Piece, PieceType};
+use crate::states::Assets;
 use crate::util::DurationExt;
+use ggez::graphics::Point;
+use ggez::{graphics, Context};
+use rand;
+use std::time::Duration;
 
 const SPEED: f32 = 170.0;
 
@@ -110,10 +110,7 @@ impl Spawner {
         // spawn at a random x position way above viewable screen so pieces just
         // dont 'pop' in but kinda just float into view
         let position = Point::new(x, -200.0);
-        self.active_pieces.push(SpawnedPiece::new(
-            position,
-            extra_rotation,
-            piece_type,
-        ));
+        self.active_pieces
+            .push(SpawnedPiece::new(position, extra_rotation, piece_type));
     }
 }
