+++ src/util.rs
@@ -17,8 +17,8 @@ impl DurationExt for Duration {
 
 #[cfg(test)]
 mod tests {
-    use std::time::Duration;
     use super::DurationExt;
+    use std::time::Duration;
 
     #[test]
     fn it_works() {
@@ -28,9 +28,8 @@ mod tests {
     }
 }
 
-
+use crate::states::Assets;
 use ggez::GameResult;
-use crate::event::Assets;
 
 /// Play the 'click' sound effect. This is a general utility method as it is
 /// used in a couple of different game states.
